requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
celery==5.4.0
redis==5.0.8
